Break up launcher into platform apps
