+++ app-launcher/android/src/main/AndroidManifest.xml
