+++ app-launcher/android/src/main/ic_launcher-web.png
