+++ app-launcher/android/src/main/kotlin/io/ashdavies/playground/LauncherActivity.kt
@@ -1,14 +1,10 @@
 package io.ashdavies.playground
 
-import android.os.Bundle
-import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.core.view.WindowCompat.setDecorFitsSystemWindows
 import com.arkivanov.decompose.defaultComponentContext
 
 private const val INTENT_EXTRA_ROUTE = "route"
-
-private typealias OnCreateAction = ComponentActivity.(savedInstanceState: Bundle?) -> Unit
 
 internal class LauncherActivity : KotlinActivity({
     val route = enumValueOfOrNull<LauncherRoute>(intent.getStringExtra(INTENT_EXTRA_ROUTE))
@@ -16,13 +12,6 @@ internal class LauncherActivity : KotlinActivity({
     setDecorFitsSystemWindows(window, true)
     setContent { LauncherApp(defaultComponentContext(), route) }
 })
-
-internal abstract class KotlinActivity(private val action: OnCreateAction) : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        action(savedInstanceState)
-    }
-}
 
 private inline fun <reified T : Enum<T>> enumValueOfOrNull(value: String?): T? {
     val term = value?.lowercase() ?: return null
