+++ app-launcher/android/src/main/res/values/strings.xml
