+++ app-launcher/common/src/commonMain/kotlin/io/ashdavies/playground/LauncherRoute.kt
@@ -1,6 +1,6 @@
 package io.ashdavies.playground
 
-internal enum class LauncherRoute {
+public enum class LauncherRoute {
     Default,
     Dominion,
     Events,
