+++ auth-oauth/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("com.android.library")
+    id("io.ashdavies.android")
     id("io.ashdavies.kotlin")
 }
