+++ build-plugins/src/main/kotlin/Project.kt
@@ -1,6 +1,8 @@
 import org.gradle.accessors.dm.LibrariesForLibs
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.getByName
+import org.jetbrains.compose.ComposePlugin
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 
 /*
  * Make version catalogs accessible from precompiled script plugins
@@ -8,3 +10,6 @@ import org.gradle.kotlin.dsl.getByName
  */
 val Project.libs: LibrariesForLibs
     get() = extensions.getByName<LibrariesForLibs>("libs")
+
+val KotlinMultiplatformExtension.compose: ComposePlugin.Dependencies
+    get() = ComposePlugin.Dependencies
