+++ build-plugins/src/main/kotlin/io.ashdavies.default.gradle.kts
@@ -1,0 +1,5 @@
+plugins {
+    // id("com.android.library")
+    // id("io.ashdavies.android")
+    // id("io.ashdavies.kotlin")
+}
