+++ build-plugins/src/main/kotlin/io.ashdavies.kotlin.gradle.kts
@@ -1,16 +1,5 @@
-import com.android.build.api.dsl.CommonExtension
-import com.android.build.gradle.LibraryExtension
-import com.android.build.gradle.internal.dsl.BaseAppModuleExtension
-import org.gradle.api.JavaVersion
-import org.gradle.kotlin.dsl.getValue
-import org.gradle.kotlin.dsl.getting
-import org.jetbrains.compose.ComposePlugin
 import org.jetbrains.compose.ExperimentalComposeLibrary
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
-
-pluginManager.withPlugin("android") {
-    plugins { id("kotlin-parcelize") }
-}
 
 plugins {
     id("org.jetbrains.compose")
@@ -21,14 +10,11 @@ plugins {
 
 kotlin {
     explicitApiWarning()
-    android()
     jvm()
 
     sourceSets.all {
         languageSettings.optIn("kotlin.RequiresOptIn")
     }
-
-    val compose = ComposePlugin.Dependencies
 
     @Suppress("UNUSED_VARIABLE")
     @OptIn(ExperimentalComposeLibrary::class)
@@ -50,47 +36,9 @@ kotlin {
     }
 
     @Suppress("UNUSED_VARIABLE")
-    val androidMain by sourceSets.dependencies {
-        implementation(libs.androidx.annotation)
-        implementation(libs.androidx.core.ktx)
-        implementation(libs.google.android.material)
-        implementation(libs.jetbrains.kotlinx.coroutines.android)
-    }
-
-    @Suppress("UNUSED_VARIABLE")
-    val androidTest by sourceSets.getting {
-        val androidAndroidTestDebug by sourceSets.getting
-        dependsOn(androidAndroidTestDebug)
-    }
-
-    @Suppress("UNUSED_VARIABLE")
     val jvmMain by sourceSets.dependencies {
         implementation(compose.desktop.currentOs)
         implementation(libs.jetbrains.kotlinx.coroutines.swing)
-    }
-}
-
-extensions.withExtension<BaseAppModuleExtension> { configure() }
-
-extensions.withExtension<LibraryExtension> { configure() }
-
-fun CommonExtension<*, *, *, *>.configure() {
-    buildFeatures {
-        compose = true
-    }
-
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-
-    defaultConfig {
-        compileSdk = 33
-        minSdk = 21
-    }
-
-    sourceSets.configureEach {
-        manifest.srcFile("src/androidMain/AndroidManifest.xml")
     }
 }
