+++ dominion-app/build.gradle.kts
@@ -1,6 +1,8 @@
 plugins {
     id("com.android.library")
+    id("io.ashdavies.android")
     id("io.ashdavies.kotlin")
+    id("kotlin-parcelize")
 }
 
 android {
