+++ events-app/build.gradle.kts
@@ -2,7 +2,9 @@ import com.android.build.api.dsl.VariantDimension
 
 plugins {
     id("com.android.library")
+    id("io.ashdavies.android")
     id("io.ashdavies.kotlin")
+    id("kotlin-parcelize")
 }
 
 android {
