+++ local-remote/build.gradle.kts
@@ -2,6 +2,7 @@ import com.android.build.api.dsl.VariantDimension
 
 plugins {
     id("com.android.library")
+    id("io.ashdavies.android")
     id("io.ashdavies.kotlin")
 }
