+++ playground-app/build.gradle.kts
@@ -1,10 +1,11 @@
 plugins {
     id("com.android.library")
+    id("io.ashdavies.android")
     id("io.ashdavies.kotlin")
 }
 
 android {
-    namespace = "io.ashdavies.common"
+    namespace = "io.ashdavies.base"
 }
 
 kotlin {
